Create core package and clean up configuration classes
